Remove CustomScope from Kafka module
