+++ gradle.properties
@@ -1,7 +1,7 @@
-projectVersion=3.3.4-SNAPSHOT
+projectVersion=4.0.0-SNAPSHOT
 micronautDocsVersion=2.0.0.RC1
 micronautBuildVersion=1.1.5
-micronautVersion=2.4.2
+micronautVersion=2.5.6
 micronautGradlePluginVersion=1.5.0
 groovyVersion=3.0.5
 spockVersion=2.0-M3-groovy-3.0
